# shared/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: py_ai_apis_gpt Gateway API
  version: 0.2.0
servers:
  - url: http://127.0.0.1:8080
security:
  - bearerAuth: []
paths:
...
  /v1/chat/completions:
    post:
      summary: OpenAI-compatible chat completions with optional streaming and RAG.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: Chat completion or SSE stream.
...
  /rag/documents/{id}:
    delete:
      summary: Delete a per-user RAG document from DB and Qdrant by stable document_id.
      parameters:
        - $ref: "#/components/parameters/IdPath"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    ChatRequest:
      type: object
      required: [messages]
      properties:
        model:
          type: string
        stream:
          type: boolean
        session_id:
          type: string
        use_rag:
          type: boolean
        rag_collection:
          type: string
        messages:
          type: array
          items:
            type: object
            required: [role, content]
            properties:
              role:
                type: string
              content:
                type: string
...